Move luaLoggerBinding.c to correct directory
